+++ packages/contracts-communication/test/mocks/SynapseGasOracleMock.sol
@@ -10,7 +10,7 @@ contract SynapseGasOracleMock is ISynapseGasOracle {
     function getLocalGasData() external view returns (bytes memory) {}
 
     function convertRemoteValueToLocalUnits(
-        uint256 remoteChainId,
+        uint64 remoteChainId,
         uint256 value
     )
         external
@@ -20,7 +20,7 @@ contract SynapseGasOracleMock is ISynapseGasOracle {
     {}
 
     function estimateTxCostInLocalUnits(
-        uint256 remoteChainId,
+        uint64 remoteChainId,
         uint256 gasLimit,
         uint256 calldataSize
     )
@@ -31,7 +31,7 @@ contract SynapseGasOracleMock is ISynapseGasOracle {
     {}
 
     function estimateTxCostInRemoteUnits(
-        uint256 remoteChainId,
+        uint64 remoteChainId,
         uint256 gasLimit,
         uint256 calldataSize
     )
